Use diff in md5 checksums to detect file changes
Problem: only the changes to the files known to the existing files
are detected. If there are new files to the directories, for
example, this is ignroed.

Solution: compute checksums in the directories specified before
and after the build, and compare the result, rather than
using `md5deep -x`.
The output in `diff` is not very friendly, so we need to
massage it with `awk`.