Inerpolate workspace/cache name
The signed URL is passed through `Shellwords.escape` before
sent to `casher`. This means that if the user gives
`name: $VAR`, the literal `$VAR` is passed to `casher`.

Ideally `travis-build` can deal with this problem, but it is
a little more involved.

So we interpolate the URLs on the fly here.